Added option to import XSL files
XSL Files can be uploaded as zip on the wiki and the same filename if
specified in the xsl_import option would be uploaded to the docbook
processing server.
Note: This can be a security risk if allowed on public wikis

Another option xsl_import_path is added which lets you include XSL from
any location on the docbook processing server. The files could be either
uploaded through the xsl_import option or through any other mechanism on
the server. This may mean that the standard docbook XSL won't be
included, however the styling defined via #docbook function would still
be applied and also override the xsl files imported.

Change-Id: Ia713e5cda692f064fd64023f81dc677cf0aad477